requirements.txt
streamlit
pandas
pybit
google-genai
python-dotenv
edge-tts